move test_transform_list_filters_to_tuple to test_utils.py
